+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.x] - 2025-12-xx - unreleased
+
+### Fixed
+
+- **DaisySelect**: Fixed dropdown popup positioning in Browser/WebAssembly when `SelectedIndex` is set during initialization
+- **run_browser.ps1**: fixed to run with debug configuration by default
+
 ## [1.4.1] - 2025-12-13
 
 ### Added
@@ -247,7 +254,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom controls: ComponentSidebar, ModifierKeys
 - Gallery demo application
 
-[1.4.0]: https://github.com/tobitege/Flowery.NET/compare/v1.3.1...HEAD
+[1.4.1]: https://github.com/tobitege/Flowery.NET/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/tobitege/Flowery.NET/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/tobitege/Flowery.NET/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/tobitege/Flowery.NET/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/tobitege/Flowery.NET/compare/v1.1.0...v1.2.0